split off typemap into its own file
