+++ src/alloc.c
@@ -740,17 +740,6 @@ jl_sym_t *jl_demangle_typename(jl_sym_t *s)
     return jl_symbol_n(&n[1], len);
 }
 
-jl_typemap_level_t *jl_new_typemap_level(void)
-{
-    jl_typemap_level_t *cache = (jl_typemap_level_t*)jl_gc_allocobj(sizeof(jl_typemap_level_t));
-    jl_set_typeof(cache, jl_typemap_level_type);
-    cache->key = NULL;
-    cache->linear = (jl_typemap_entry_t*)jl_nothing;
-    cache->arg1 = (jl_array_t*)jl_nothing;
-    cache->targ = (jl_array_t*)jl_nothing;
-    return cache;
-}
-
 JL_DLLEXPORT jl_methtable_t *jl_new_method_table(jl_sym_t *name, jl_module_t *module)
 {
     jl_methtable_t *mt = (jl_methtable_t*)jl_gc_allocobj(sizeof(jl_methtable_t));
